Change to higher resolution timer

+++ FerretLib.SFML/ChronoEventArgs.cs
@@ -4,14 +4,14 @@ namespace FerretLib.SFML
 {
     public class ChronoEventArgs : EventArgs
     {
-        public ChronoEventArgs(int monotonic, double delta, double fps)
+        public ChronoEventArgs(long monotonic, double delta, double fps)
         {
             Monotonic = monotonic;
             Delta = delta;
             Fps = fps;
         }
 
-        public int Monotonic { get; private set; }
+        public long Monotonic { get; private set; }
         public double Delta { get; private set; }
         public double Fps { get; private set; }
     }
